Fix tests, add Makefile, add travis
