+++ .gitignore
@@ -57,3 +57,5 @@ docs/_build/
 
 # PyBuilder
 target/
+
+test.db
