+++ .travis.yml
@@ -1,0 +1,26 @@
+language: python
+
+python:
+  - 2.7
+  - 3.3
+  - 3.4
+  - 3.5
+
+before_install:
+  - pip install coveralls
+
+install:
+  - pip install -e .
+  - pip install -r requirements.txt
+  - pip install Django==${DJANGO}
+
+script: coverage run --source chartit runtests.py chartit_tests
+
+env:
+  - DJANGO=1.8.8
+  - DJANGO=1.9.1
+
+matrix:
+  exclude:
+    - python: 3.3
+      env: DJANGO=1.9.1
