+++ chartit_tests/tests.py
@@ -2,8 +2,7 @@ from django.test import TestCase
 from django.db.models import Avg
 
 from chartit import PivotDataPool, DataPool
-from chartit.validation import (clean_pdps, clean_dps,
-                                clean_pcso, clean_cso)
+from chartit.validation import clean_pdps, clean_dps, clean_pcso, clean_cso
 from chartit.exceptions import APIInputError
 
 from .models import SalesHistory, MonthlyWeatherByCity, MonthlyWeatherSeattle
@@ -11,15 +10,16 @@ from .utils import assertOptionDictsEqual
 
 TestCase.assertOptionDictsEqual = assertOptionDictsEqual
 
-class GoodPivotSeriesDictInput(TestCase):
-    
+
+class GoodPivotSeriesDictInputTests(TestCase):
+
     def test_all_terms(self):
         series_input = \
-          {'avg_price': { 
+          {'avg_price': {
              'source': SalesHistory.objects.all(),
              'func': Avg('price'),
              'categories': [
-               'bookstore__city__state', 
+               'bookstore__city__state',
                'bookstore__city__city'],
              'legend_by': ['book__genre__name'],
              'top_n_per_cat': 5,
@@ -28,11 +28,11 @@ class GoodPivotSeriesDictInput(TestCase):
                'bookstore__city__city': 'city',
                'book__genre__name': 'name'}}}
         series_cleaned = \
-          {'avg_price': { 
+          {'avg_price': {
              'source': SalesHistory.objects.all(),
              'func': Avg('price'),
              'categories': [
-               'bookstore__city__state', 
+               'bookstore__city__state',
                'bookstore__city__city'],
              'legend_by': ['book__genre__name'],
              'top_n_per_cat': 5,
@@ -41,11 +41,11 @@ class GoodPivotSeriesDictInput(TestCase):
                'bookstore__city__city': 'city',
                'book__genre__name': 'name'}}}
         self.assertOptionDictsEqual(clean_pdps(series_input),
-                                   series_cleaned)
-    
+                                    series_cleaned)
+
     def test_categories_is_a_str(self):
         series_input = \
-          {'avg_price': { 
+          {'avg_price': {
              'source': SalesHistory.objects.all(),
              'func': Avg('price'),
              'categories': 'bookstore__city__state',
@@ -55,7 +55,7 @@ class GoodPivotSeriesDictInput(TestCase):
                'bookstore__city__state': 'state',
                'book__genre__name': 'name'}}}
         series_cleaned = \
-          {'avg_price': { 
+          {'avg_price': {
              'source': SalesHistory.objects.all(),
              'func': Avg('price'),
              'categories': ['bookstore__city__state'],
@@ -65,15 +65,15 @@ class GoodPivotSeriesDictInput(TestCase):
                'bookstore__city__state': 'state',
                'book__genre__name': 'name'}}}
         self.assertOptionDictsEqual(clean_pdps(series_input),
-                                   series_cleaned)
-    
+                                    series_cleaned)
+
     def test_legend_by_is_a_str(self):
         series_input = \
-          {'avg_price': { 
+          {'avg_price': {
              'source': SalesHistory.objects.all(),
              'func': Avg('price'),
              'categories': [
-               'bookstore__city__state', 
+               'bookstore__city__state',
                'bookstore__city__city'],
              'legend_by': 'book__genre__name',
              'top_n_per_cat': 5,
@@ -82,11 +82,11 @@ class GoodPivotSeriesDictInput(TestCase):
                'bookstore__city__city': 'city',
                'book__genre__name': 'name'}}}
         series_cleaned = \
-          {'avg_price': { 
+          {'avg_price': {
              'source': SalesHistory.objects.all(),
              'func': Avg('price'),
              'categories': [
-               'bookstore__city__state', 
+               'bookstore__city__state',
                'bookstore__city__city'],
              'legend_by': ['book__genre__name'],
              'top_n_per_cat': 5,
@@ -95,26 +95,26 @@ class GoodPivotSeriesDictInput(TestCase):
                'bookstore__city__city': 'city',
                'book__genre__name': 'name'}}}
         self.assertOptionDictsEqual(clean_pdps(series_input),
-                                   series_cleaned)
-        
+                                    series_cleaned)
+
     def test_no_legend_by(self):
         series_input = \
-          {'avg_price': { 
+          {'avg_price': {
              'source': SalesHistory.objects.all(),
              'func': Avg('price'),
              'categories': [
-               'bookstore__city__state', 
+               'bookstore__city__state',
                'bookstore__city__city'],
              'top_n_per_cat': 5,
              'field_aliases': {
                'bookstore__city__state': 'state',
                'bookstore__city__city': 'city'}}}
         series_cleaned = \
-          {'avg_price': { 
+          {'avg_price': {
              'source': SalesHistory.objects.all(),
              'func': Avg('price'),
              'categories': [
-               'bookstore__city__state', 
+               'bookstore__city__state',
                'bookstore__city__city'],
              'legend_by': (),
              'top_n_per_cat': 5,
@@ -123,10 +123,10 @@ class GoodPivotSeriesDictInput(TestCase):
                'bookstore__city__city': 'city'}}}
         self.assertOptionDictsEqual(clean_pdps(series_input),
                                     series_cleaned)
-        
+
     def test_no_top_n_per_cat(self):
         series_input = \
-          {'avg_price': { 
+          {'avg_price': {
              'source': SalesHistory.objects.all(),
              'func': Avg('price'),
              'categories': ['bookstore__city__state', 'bookstore__city__city'],
@@ -136,7 +136,7 @@ class GoodPivotSeriesDictInput(TestCase):
                'bookstore__city__city': 'city',
                'book__genre__name': 'name'}}}
         series_cleaned = \
-          {'avg_price': { 
+          {'avg_price': {
              'source': SalesHistory.objects.all(),
              'func': Avg('price'),
              'categories': ['bookstore__city__state', 'bookstore__city__city'],
@@ -146,20 +146,20 @@ class GoodPivotSeriesDictInput(TestCase):
                'bookstore__city__state': 'state',
                'bookstore__city__city': 'city',
                'book__genre__name': 'name'}}}
-        
+
         self.assertOptionDictsEqual(clean_pdps(series_input),
-                                   series_cleaned)
-    
+                                    series_cleaned)
+
     def test_no_field_aliases(self):
         series_input = \
-          {'avg_price': { 
+          {'avg_price': {
              'source': SalesHistory.objects.all(),
              'func': Avg('price'),
              'categories': ['bookstore__city__state', 'bookstore__city__city'],
              'legend_by': ['book__genre__name'],
              'top_n_per_cat': 5}}
         series_cleaned = \
-          {'avg_price': { 
+          {'avg_price': {
             'source': SalesHistory.objects.all(),
             'func': Avg('price'),
             'categories': ['bookstore__city__state', 'bookstore__city__city'],
@@ -170,11 +170,11 @@ class GoodPivotSeriesDictInput(TestCase):
               'bookstore__city__city': 'city',
               'book__genre__name': 'name'}}}
         self.assertOptionDictsEqual(clean_pdps(series_input),
-                                   series_cleaned)
-        
+                                    series_cleaned)
+
     def test_custom_field_aliases(self):
         series_input = \
-          {'avg_price': { 
+          {'avg_price': {
              'source': SalesHistory.objects.all(),
              'func': Avg('price'),
              'categories': ['bookstore__city__state', 'bookstore__city__city'],
@@ -185,7 +185,7 @@ class GoodPivotSeriesDictInput(TestCase):
                'bookstore__city__city': 'Cty',
                'book__genre__name': 'Genre'}}}
         series_cleaned = \
-          {'avg_price': { 
+          {'avg_price': {
              'source': SalesHistory.objects.all(),
              'func': Avg('price'),
              'categories': ['bookstore__city__state', 'bookstore__city__city'],
@@ -196,11 +196,11 @@ class GoodPivotSeriesDictInput(TestCase):
                'bookstore__city__city': 'Cty',
                'book__genre__name': 'Genre'}}}
         self.assertOptionDictsEqual(clean_pdps(series_input),
-                                   series_cleaned) 
-        
+                                    series_cleaned)
+
     def test_partial_field_aliases(self):
         series_input = \
-          {'avg_price': { 
+          {'avg_price': {
              'source': SalesHistory.objects.all(),
              'func': Avg('price'),
              'categories': ['bookstore__city__state', 'bookstore__city__city'],
@@ -209,7 +209,7 @@ class GoodPivotSeriesDictInput(TestCase):
              'field_aliases': {
                'bookstore__city__state': 'St'}}}
         series_cleaned = \
-          {'avg_price': { 
+          {'avg_price': {
              'source': SalesHistory.objects.all(),
              'func': Avg('price'),
              'categories': ['bookstore__city__state', 'bookstore__city__city'],
@@ -220,23 +220,22 @@ class GoodPivotSeriesDictInput(TestCase):
                'bookstore__city__city': 'city',
                'book__genre__name': 'name'}}}
         self.assertOptionDictsEqual(clean_pdps(series_input),
-                                   series_cleaned)
-        
+                                    series_cleaned)
 
-class BadPivotSeriesDictInput(TestCase):
-    
+
+class BadPivotSeriesDictInputTests(TestCase):
+
     def test_series_not_dict_or_list(self):
         series_input = 'foobar'
         self.assertRaises(APIInputError, clean_pdps, series_input)
-        
+
     def test_func_dict_wrong_type(self):
-        series_input = \
-          {'avg_price': 'foobar'}
+        series_input = {'avg_price': 'foobar'}
         self.assertRaises(APIInputError, clean_pdps, series_input)
-        
+
     def test_source_missing(self):
         series_input = \
-          {'avg_price': { 
+          {'avg_price': {
              'func': Avg('price'),
              'categories': ['bookstore__city__state', 'bookstore__city__city'],
              'legend_by': ['book__genre__name'],
@@ -246,10 +245,10 @@ class BadPivotSeriesDictInput(TestCase):
                'bookstore__city__city': 'city',
                'book__genre__name': 'name'}}}
         self.assertRaises(APIInputError, clean_pdps, series_input)
-    
+
     def test_source_wrong_type(self):
         series_input = \
-          {'avg_price': { 
+          {'avg_price': {
              'source': 'foobar',
              'func': Avg('price'),
              'categories': ['bookstore__city__state', 'bookstore__city__city'],
@@ -260,10 +259,10 @@ class BadPivotSeriesDictInput(TestCase):
                'bookstore__city__city': 'city',
                'book__genre__name': 'name'}}}
         self.assertRaises(APIInputError, clean_pdps, series_input)
-        
+
     def test_func_missing(self):
         series_input = \
-          {'avg_price': { 
+          {'avg_price': {
              'source': SalesHistory.objects.all(),
              'categories': ['bookstore__city__state', 'bookstore__city__city'],
              'legend_by': ['book__genre__name'],
@@ -273,10 +272,10 @@ class BadPivotSeriesDictInput(TestCase):
                'bookstore__city__city': 'city',
                'book__genre__name': 'name'}}}
         self.assertRaises(APIInputError, clean_pdps, series_input)
-        
+
     def test_func_wrong_type(self):
         series_input = \
-          {'avg_price': { 
+          {'avg_price': {
              'source': SalesHistory.objects.all(),
              'func': 'foobar',
              'categories': ['bookstore__city__state', 'bookstore__city__city'],
@@ -287,10 +286,10 @@ class BadPivotSeriesDictInput(TestCase):
                'bookstore__city__city': 'city',
                'book__genre__name': 'name'}}}
         self.assertRaises(APIInputError, clean_pdps, series_input)
-        
+
     def test_categories_missing(self):
         series_input = \
-          {'avg_price': { 
+          {'avg_price': {
              'source': SalesHistory.objects.all(),
              'func': Avg('price'),
              'legend_by': ['book__genre__name'],
@@ -300,10 +299,10 @@ class BadPivotSeriesDictInput(TestCase):
                'bookstore__city__city': 'city',
                'book__genre__name': 'name'}}}
         self.assertRaises(APIInputError, clean_pdps, series_input)
-        
+
     def test_categories_wrong_type(self):
         series_input = \
-          {'avg_price': { 
+          {'avg_price': {
              'source': SalesHistory.objects.all(),
              'func': Avg('price'),
              'categories': 0,
@@ -314,10 +313,10 @@ class BadPivotSeriesDictInput(TestCase):
                'bookstore__city__city': 'city',
                'book__genre__name': 'name'}}}
         self.assertRaises(APIInputError, clean_pdps, series_input)
-        
+
     def test_categories_not_a_valid_field(self):
         series_input = \
-          {'avg_price': { 
+          {'avg_price': {
              'source': SalesHistory.objects.all(),
              'func': Avg('price'),
              'categories': ['foobar'],
@@ -331,7 +330,7 @@ class BadPivotSeriesDictInput(TestCase):
 
     def test_categories_empty_list(self):
         series_input = \
-          {'avg_price': { 
+          {'avg_price': {
              'source': SalesHistory.objects.all(),
              'func': Avg('price'),
              'categories': [],
@@ -342,10 +341,10 @@ class BadPivotSeriesDictInput(TestCase):
                'bookstore__city__city': 'city',
                'book__genre__name': 'name'}}}
         self.assertRaises(APIInputError, clean_pdps, series_input)
-                
+
     def test_legend_by_wrong_type(self):
         series_input = \
-          {'avg_price': { 
+          {'avg_price': {
              'source': SalesHistory.objects.all(),
              'func': 'foobar',
              'categories': ['bookstore__city__state', 'bookstore__city__city'],
@@ -356,10 +355,10 @@ class BadPivotSeriesDictInput(TestCase):
                'bookstore__city__city': 'city',
                'book__genre__name': 'name'}}}
         self.assertRaises(APIInputError, clean_pdps, series_input)
-        
+
     def test_legend_by_not_a_valid_field(self):
         series_input = \
-          {'avg_price': { 
+          {'avg_price': {
              'source': SalesHistory.objects.all(),
              'func': 'foobar',
              'categories': ['bookstore__city__state', 'bookstore__city__city'],
@@ -369,11 +368,11 @@ class BadPivotSeriesDictInput(TestCase):
                'bookstore__city__state': 'state',
                'bookstore__city__city': 'city',
                'book__genre__name': 'name'}}}
-        self.assertRaises(APIInputError, clean_pdps, series_input)    
-    
+        self.assertRaises(APIInputError, clean_pdps, series_input)
+
     def test_top_n_per_cat_wrong_type(self):
         series_input = \
-          {'avg_price': { 
+          {'avg_price': {
              'source': SalesHistory.objects.all(),
              'func': 'foobar',
              'categories': ['bookstore__city__state', 'bookstore__city__city'],
@@ -384,13 +383,13 @@ class BadPivotSeriesDictInput(TestCase):
                'bookstore__city__city': 'city',
                'book__genre__name': 'name'}}}
         self.assertRaises(APIInputError, clean_pdps, series_input)
-    
-class GoodPivotSeriesListInput(TestCase):  
-    
+
+
+class GoodPivotSeriesListInputTests(TestCase):
+
     def test_all_terms(self):
-                
         series_input = \
-          [{'options': 
+          [{'options':
              {'source': SalesHistory.objects.all(),
               'categories': 'bookstore__city__state',
               'legend_by': 'book__genre__name',
@@ -419,14 +418,13 @@ class GoodPivotSeriesListInput(TestCase):
              'top_n_per_cat': 2,
              'field_aliases': {
                'bookstore__city__state': 'state'}}}
-        
+
         self.assertOptionDictsEqual(clean_pdps(series_input),
-                                   series_cleaned)
+                                    series_cleaned)
 
     def test_source_a_manager(self):
-                
         series_input = \
-          [{'options': 
+          [{'options':
              {'source': SalesHistory.objects,
               'categories': 'bookstore__city__state',
               'legend_by': 'book__genre__name',
@@ -455,14 +453,13 @@ class GoodPivotSeriesListInput(TestCase):
              'top_n_per_cat': 2,
              'field_aliases': {
                'bookstore__city__state': 'state'}}}
-        
+
         self.assertOptionDictsEqual(clean_pdps(series_input),
-                                   series_cleaned)
-    
+                                    series_cleaned)
+
     def test_source_a_model(self):
-                
         series_input = \
-          [{'options': 
+          [{'options':
              {'source': SalesHistory,
               'categories': 'bookstore__city__state',
               'legend_by': 'book__genre__name',
@@ -491,14 +488,13 @@ class GoodPivotSeriesListInput(TestCase):
              'top_n_per_cat': 2,
              'field_aliases': {
                'bookstore__city__state': 'state'}}}
-        
+
         self.assertOptionDictsEqual(clean_pdps(series_input),
-                                   series_cleaned)
-         
+                                    series_cleaned)
+
     def test_term_opts_an_aggr(self):
-        
         series_input = \
-          [{'options': 
+          [{'options':
              {'source': SalesHistory.objects.all(),
               'categories': ['bookstore__city__state'],
               'legend_by': ['book__genre__name'],
@@ -516,15 +512,13 @@ class GoodPivotSeriesListInput(TestCase):
             'field_aliases': {
               'bookstore__city__state': 'state',
               'book__genre__name': 'name'}}}
-        
-        self.assertOptionDictsEqual(clean_pdps(series_input),
-                                   series_cleaned)
 
+        self.assertOptionDictsEqual(clean_pdps(series_input),
+                                    series_cleaned)
 
     def test_term_opts_a_dict(self):
-        
         series_input = \
-          [{'options': 
+          [{'options':
              {'source': SalesHistory.objects.all(),
               'categories': 'bookstore__city__state',
               'legend_by': 'book__genre__name',
@@ -544,12 +538,11 @@ class GoodPivotSeriesListInput(TestCase):
             'field_aliases': {
               'bookstore__city__state': 'state',
               'book__genre__name': 'name'}}}
-        
+
         self.assertOptionDictsEqual(clean_pdps(series_input),
-                                   series_cleaned)
-    
+                                    series_cleaned)
+
     def test_opts_empty(self):
-        
         series_input = \
           [{'options': {},
             'terms': {
@@ -568,12 +561,11 @@ class GoodPivotSeriesListInput(TestCase):
             'top_n_per_cat': 3,
             'field_aliases': {
               'bookstore__city__state': 'state'}}}
-        
+
         self.assertOptionDictsEqual(clean_pdps(series_input),
-                                   series_cleaned)
+                                    series_cleaned)
 
     def test_categories_a_str(self):
-        
         series_input = \
           [{'options': {},
             'terms': {
@@ -592,14 +584,13 @@ class GoodPivotSeriesListInput(TestCase):
             'top_n_per_cat': 3,
             'field_aliases': {
               'bookstore__city__state': 'state'}}}
-        
+
         self.assertOptionDictsEqual(clean_pdps(series_input),
-                                   series_cleaned)
+                                    series_cleaned)
 
     def test_legend_by_a_str(self):
-        
         series_input = \
-          [{'options': 
+          [{'options':
              {'source': SalesHistory.objects.all(),
               'categories': ['bookstore__city__state'],
               'legend_by': 'book__genre__name',
@@ -617,21 +608,20 @@ class GoodPivotSeriesListInput(TestCase):
             'field_aliases': {
               'bookstore__city__state': 'state',
               'book__genre__name': 'name'}}}
-        
+
         self.assertOptionDictsEqual(clean_pdps(series_input),
-                                   series_cleaned)
+                                    series_cleaned)
 
     def test_multiple_dicts(self):
-                
         series_input = \
-          [{'options': 
+          [{'options':
              {'source': SalesHistory.objects.all(),
               'categories': 'bookstore__city__state',
               'legend_by': 'book__genre__name',
               'top_n_per_cat': 2},
             'terms': {
               'avg_price': Avg('price')}},
-           {'options': 
+           {'options':
              {'source': SalesHistory.objects.filter(price__gte=10),
               'categories': 'bookstore__city__city',
               'top_n_per_cat': 2},
@@ -658,36 +648,36 @@ class GoodPivotSeriesListInput(TestCase):
              'top_n_per_cat': 2,
              'field_aliases': {
                'bookstore__city__city': 'city'}}}
-        
+
         self.assertOptionDictsEqual(clean_pdps(series_input),
-                                   series_cleaned)
-        
-class BadPivotSeriesListInput(TestCase):  
-    
+                                    series_cleaned)
+
+
+class BadPivotSeriesListInputTests(TestCase):
+
     def test_terms_empty(self):
-                
         series_input = \
-          [{'options': 
+          [{'options':
              {'source': SalesHistory.objects.all(),
               'categories': 'bookstore__city__state',
               'legend_by': 'book__genre__name',
               'top_n_per_cat': 2},
             'terms': {}}]
-          
+
         self.assertRaises(APIInputError, clean_pdps, series_input)
-        
+
     def test_terms_missing(self):
         series_input = \
-          [{'options': 
+          [{'options':
              {'source': SalesHistory.objects.all(),
               'categories': 'bookstore__city__state',
               'legend_by': 'book__genre__name',
               'top_n_per_cat': 2}}]
         self.assertRaises(APIInputError, clean_pdps, series_input)
-    
+
     def test_terms_a_list_not_a_dict(self):
         series_input = \
-          [{'options': 
+          [{'options':
              {'source': SalesHistory.objects.all(),
               'categories': 'bookstore__city__state',
               'legend_by': 'book__genre__name',
@@ -698,10 +688,10 @@ class BadPivotSeriesListInput(TestCase):
                 'func': Avg('price'),
                 'legend_by': None}}]}]
         self.assertRaises(APIInputError, clean_pdps, series_input)
-        
+
     def test_source_missing(self):
         series_input = \
-          [{'options': 
+          [{'options':
              {'categories': 'bookstore__city__state',
               'legend_by': 'book__genre__name',
               'top_n_per_cat': 2},
@@ -711,7 +701,7 @@ class BadPivotSeriesListInput(TestCase):
                 'func': Avg('price'),
                 'legend_by': None}}}]
         self.assertRaises(APIInputError, clean_pdps, series_input)
-    
+
     def test_options_missing(self):
         series_input = \
           [{'terms': {
@@ -719,8 +709,8 @@ class BadPivotSeriesListInput(TestCase):
               'avg_price_all': {
                 'func': Avg('price'),
                 'legend_by': None}}}]
-        self.assertRaises(APIInputError, clean_pdps, series_input)  
-    
+        self.assertRaises(APIInputError, clean_pdps, series_input)
+
     def test_options_empty(self):
         series_input = \
           [{'options': {},
@@ -729,11 +719,11 @@ class BadPivotSeriesListInput(TestCase):
               'avg_price_all': {
                 'func': Avg('price'),
                 'legend_by': None}}}]
-        self.assertRaises(APIInputError, clean_pdps, series_input)  
-        
+        self.assertRaises(APIInputError, clean_pdps, series_input)
+
     def test_source_wrong_type(self):
         series_input = \
-          [{'options': 
+          [{'options':
              {'source': 'foobar',
               'categories': 'bookstore__city__state',
               'legend_by': 'book__genre__name',
@@ -743,11 +733,11 @@ class BadPivotSeriesListInput(TestCase):
               'avg_price_all': {
                 'func': Avg('price'),
                 'legend_by': None}}}]
-        self.assertRaises(APIInputError, clean_pdps, series_input)  
-    
+        self.assertRaises(APIInputError, clean_pdps, series_input)
+
     def test_categories_wrong_type(self):
         series_input = \
-          [{'options': 
+          [{'options':
              {'source': SalesHistory.objects.all(),
               'categories': 10,
               'legend_by': 'book__genre__name',
@@ -758,10 +748,10 @@ class BadPivotSeriesListInput(TestCase):
                 'func': Avg('price'),
                 'legend_by': None}}}]
         self.assertRaises(APIInputError, clean_pdps, series_input)
-    
+
     def test_categories_not_a_field(self):
         series_input = \
-          [{'options': 
+          [{'options':
              {'source': SalesHistory.objects.all(),
               'categories': 'foobar',
               'legend_by': 'book__genre__name',
@@ -772,10 +762,10 @@ class BadPivotSeriesListInput(TestCase):
                 'func': Avg('price'),
                 'legend_by': None}}}]
         self.assertRaises(APIInputError, clean_pdps, series_input)
-        
+
     def test_legend_by_wrong_type(self):
         series_input = \
-          [{'options': 
+          [{'options':
              {'source': SalesHistory.objects.all(),
               'categories': 'bookstore__city__state',
               'legend_by': 10,
@@ -789,7 +779,7 @@ class BadPivotSeriesListInput(TestCase):
 
     def test_legend_by_not_a_field(self):
         series_input = \
-          [{'options': 
+          [{'options':
              {'source': SalesHistory.objects.all(),
               'categories': 'bookstore__city__state',
               'legend_by': 'foobar',
@@ -800,10 +790,10 @@ class BadPivotSeriesListInput(TestCase):
                 'func': Avg('price'),
                 'legend_by': None}}}]
         self.assertRaises(APIInputError, clean_pdps, series_input)
-        
+
     def test_term_func_wrong_type(self):
         series_input = \
-          [{'options': 
+          [{'options':
              {'source': SalesHistory.objects.all(),
               'categories': 'bookstore__city__state',
               'legend_by': 'book__genre__name',
@@ -814,10 +804,10 @@ class BadPivotSeriesListInput(TestCase):
                 'func': Avg('price'),
                 'legend_by': None}}}]
         self.assertRaises(APIInputError, clean_pdps, series_input)
-        
+
     def test_term_dict_func_wrong_type(self):
         series_input = \
-          [{'options': 
+          [{'options':
              {'source': SalesHistory.objects.all(),
               'categories': 'bookstore__city__state',
               'legend_by': 'book__genre__name',
@@ -828,10 +818,10 @@ class BadPivotSeriesListInput(TestCase):
                 'func': 'foobar',
                 'legend_by': None}}}]
         self.assertRaises(APIInputError, clean_pdps, series_input)
-        
+
     def test_term_dict_legend_by_wrong_type(self):
         series_input = \
-          [{'options': 
+          [{'options':
              {'source': SalesHistory.objects.all(),
               'categories': 'bookstore__city__state',
               'legend_by': 'book__genre__name',
@@ -842,15 +832,15 @@ class BadPivotSeriesListInput(TestCase):
                 'func': Avg('price'),
                 'legend_by': 10}}}]
         self.assertRaises(APIInputError, clean_pdps, series_input)
-                    
-class GoodDataSeriesListInput(TestCase):
-    
+
+class GoodDataSeriesListInputTests(TestCase):
+
     def test_all_terms(self):
         series_input = \
-          [{'options': 
+          [{'options':
              {'source': SalesHistory.objects.all()},
             'terms': [
-              'price', 
+              'price',
               {'genre': {
                  'field': 'book__genre__name',
                  'source': SalesHistory.objects.filter(price__gte=10),
@@ -867,13 +857,13 @@ class GoodDataSeriesListInput(TestCase):
              'field_alias': 'gnr'}}
         self.assertOptionDictsEqual(clean_dps(series_input),
                                     series_cleaned)
-    
+
     def test_terms_list_all_str(self):
         series_input = \
-          [{'options': 
+          [{'options':
              {'source': SalesHistory.objects.all()},
             'terms': [
-              'price', 
+              'price',
               'book__genre__name']
             }]
         series_cleaned = \
@@ -887,10 +877,10 @@ class GoodDataSeriesListInput(TestCase):
              'field_alias': 'name'}}
         self.assertOptionDictsEqual(clean_dps(series_input),
                                     series_cleaned)
-    
+
     def test_terms_is_a_dict(self):
         series_input = \
-          [{'options': 
+          [{'options':
              {'source': SalesHistory.objects.all()},
             'terms': {'price': {}}
             }]
@@ -901,16 +891,16 @@ class GoodDataSeriesListInput(TestCase):
              'field_alias': 'price'}}
         self.assertOptionDictsEqual(clean_dps(series_input),
                                     series_cleaned)
-        
+
     def test_multiple_dicts(self):
         series_input = \
-          [{'options': 
+          [{'options':
              {'source': SalesHistory.objects.all()},
             'terms': [
               'price']},
-           {'options': 
+           {'options':
              {'source': SalesHistory.objects.filter(price__gte=10)},
-            'terms': 
+            'terms':
               {'genre': {
                  'field': 'book__genre__name',
                  'field_alias': 'gnr'}}
@@ -926,36 +916,36 @@ class GoodDataSeriesListInput(TestCase):
              'field_alias': 'gnr'}}
         self.assertOptionDictsEqual(clean_dps(series_input),
                                     series_cleaned)
-        
-class BadDataSeriesListInput(TestCase):
+
+class BadDataSeriesListInputTests(TestCase):
     def test_source_missing(self):
         series_input = \
           [{'options': {},
             'terms': [
-              'price', 
+              'price',
               {'genre': {
                  'field': 'book__genre__name',
                  'source': SalesHistory.objects.filter(price__gte=10),
                  'field_alias': 'gnr'}}]
             }]
         self.assertRaises(APIInputError, clean_dps, series_input)
-        
+
     def test_source_wrong_type(self):
         series_input = \
-          [{'options': 
+          [{'options':
               {'source': 'foobar'},
             'terms': [
-              'price', 
+              'price',
               {'genre': {
                'field': 'book__genre__name',
                'source': SalesHistory.objects.filter(price__gte=10),
                'field_alias': 'gnr'}}]
             }]
         self.assertRaises(APIInputError, clean_dps, series_input)
-        
+
     def test_series_terms_empty(self):
         series_input = \
-          [{'options': 
+          [{'options':
              {'source': SalesHistory.objects.all()},
             'terms': []
             }]
@@ -963,25 +953,25 @@ class BadDataSeriesListInput(TestCase):
 
     def test_series_terms_wrong_type(self):
         series_input = \
-          [{'options': 
+          [{'options':
              {'source': SalesHistory.objects.all()},
             'terms': 'foobar'
             }]
         self.assertRaises(APIInputError, clean_dps, series_input)
-        
+
     def test_terms_element_wrong_type(self):
         series_input = \
-          [{'options': 
+          [{'options':
               {'source': SalesHistory.objects.all()},
             'terms': [10]}]
         self.assertRaises(APIInputError, clean_dps, series_input)
-        
+
     def test_terms_element_not_a_field(self):
         series_input = \
-          [{'options': 
+          [{'options':
               {'source': SalesHistory.objects.all()},
             'terms': [
-              'foobar', 
+              'foobar',
               {'genre': {
                'field': 'book__genre__name',
                'source': SalesHistory.objects.filter(price__gte=10),
@@ -989,20 +979,23 @@ class BadDataSeriesListInput(TestCase):
             }]
         self.assertRaises(APIInputError, clean_dps, series_input)
 
-class GoodPivotChartOptions(TestCase):
-    series_input = \
-      [{'options': 
-         {'source': SalesHistory.objects.all(),
-          'categories': 'bookstore__city__state',
-          'legend_by': 'book__genre__name',
-          'top_n_per_cat': 2},
-        'terms': {
-          'avg_price': Avg('price'),
-          'avg_price_all': {
-            'func': Avg('price'),
-            'legend_by': None}}}]
-    ds = PivotDataPool(series_input)
-    
+
+class GoodPivotChartOptionsTests(TestCase):
+
+    def setUp(self):
+        series_input = \
+            [{'options':
+               {'source': SalesHistory.objects.all(),
+                'categories': 'bookstore__city__state',
+                'legend_by': 'book__genre__name',
+                'top_n_per_cat': 2},
+              'terms': {
+                'avg_price': Avg('price'),
+                'avg_price_all': {
+                  'func': Avg('price'),
+                  'legend_by': None}}}]
+        self.ds = PivotDataPool(series_input)
+
     def test_all_terms(self):
         pcso_input = \
           [{'options': {
@@ -1019,22 +1012,24 @@ class GoodPivotChartOptions(TestCase):
               'type': 'area'}}
         self.assertOptionDictsEqual(clean_pcso(pcso_input, self.ds),
                                     series_cleaned)
-    
 
-class BadPivotChartOptions(TestCase):
-    series_input = \
-      [{'options': 
-         {'source': SalesHistory.objects.all(),
-          'categories': 'bookstore__city__state',
-          'legend_by': 'book__genre__name',
-          'top_n_per_cat': 2},
-        'terms': {
-          'avg_price': Avg('price'),
-          'avg_price_all': {
-            'func': Avg('price'),
-            'legend_by': None}}}]
-    ds = PivotDataPool(series_input)
-    
+
+class BadPivotChartOptionsTests(TestCase):
+
+    def setUp(self):
+        series_input = \
+          [{'options':
+             {'source': SalesHistory.objects.all(),
+              'categories': 'bookstore__city__state',
+              'legend_by': 'book__genre__name',
+              'top_n_per_cat': 2},
+            'terms': {
+              'avg_price': Avg('price'),
+              'avg_price_all': {
+                'func': Avg('price'),
+                'legend_by': None}}}]
+        self.ds = PivotDataPool(series_input)
+
     def test_term_not_in_pdps(self):
         pcso_input = \
           [{'options': {
@@ -1045,7 +1040,7 @@ class BadPivotChartOptions(TestCase):
                  'type': 'area'}}]}
            ]
         self.assertRaises(APIInputError, clean_pcso, pcso_input, self.ds)
-  
+
     def test_opts_missing(self):
         pcso_input = \
           [{'terms':[
@@ -1079,7 +1074,7 @@ class BadPivotChartOptions(TestCase):
               {'avg_price_all':{
                  'type': 'area'}}}]
         self.assertRaises(APIInputError, clean_pcso, pcso_input, self.ds)
-  
+
     def test_terms_a_str(self):
         pcso_input = \
           [{'options': {
@@ -1087,25 +1082,27 @@ class BadPivotChartOptions(TestCase):
             'terms':
               'foobar'}]
         self.assertRaises(APIInputError, clean_pcso, pcso_input, self.ds)
-  
-          
-class GoodChartOptions(TestCase):
-    series_input = \
-      [{'options': {
-          'source': MonthlyWeatherByCity.objects.all()},
-        'terms': [
-          'month',
-          'boston_temp',
-          'houston_temp',
-          'new_york_temp']},
-       {'options': {
-          'source': MonthlyWeatherSeattle.objects.all()},
-        'terms': [
-          {'month_seattle': 'month'},
-          'seattle_temp']
-        }]
-    ds = DataPool(series_input) 
-    
+
+
+class GoodChartOptionsTests(TestCase):
+
+    def setUp(self):
+        series_input = \
+          [{'options': {
+              'source': MonthlyWeatherByCity.objects.all()},
+            'terms': [
+              'month',
+              'boston_temp',
+              'houston_temp',
+              'new_york_temp']},
+           {'options': {
+              'source': MonthlyWeatherSeattle.objects.all()},
+            'terms': [
+              {'month_seattle': 'month'},
+              'seattle_temp']
+            }]
+        self.ds = DataPool(series_input)
+
     def test_all_terms(self):
         so_input = \
           [{'options': {
@@ -1139,7 +1136,7 @@ class GoodChartOptions(TestCase):
               'type': 'column'},
             'terms': {
               'month':[
-                 'boston_temp', 
+                 'boston_temp',
                  'new_york_temp']}
             }]
         so_cleaned = \
@@ -1154,14 +1151,14 @@ class GoodChartOptions(TestCase):
 
     def test_all_terms_dict(self):
         so_input = \
-          [{'options': 
-             {'type': 
+          [{'options':
+             {'type':
                 'column'},
-            'terms': 
+            'terms':
               {'month':[
                  {'boston_temp': {
                     'type': 'area',
-                    'xAxis': 1}}, 
+                    'xAxis': 1}},
                  {'new_york_temp':
                     {'xAxis':0}}]}
             }]
@@ -1181,7 +1178,7 @@ class GoodChartOptions(TestCase):
         so_input = \
           [{'options':{
               'type': 'column'},
-            'terms': 
+            'terms':
               {'month':[
                  'boston_temp',
                  'new_york_temp']}
@@ -1204,26 +1201,27 @@ class GoodChartOptions(TestCase):
              'type': 'area'}}
         self.assertOptionDictsEqual(clean_cso(so_input, self.ds),
                                     so_cleaned)
-    
 
-class BadChartOptions(TestCase):
 
-    series_input = \
-      [{'options': {
-          'source': MonthlyWeatherByCity.objects.all()},
-        'terms': [
-          'month',
-          'boston_temp',
-          'houston_temp',
-          'new_york_temp']},
-       {'options': {
-          'source': MonthlyWeatherSeattle.objects.all()},
-        'terms': [
-          {'month_seattle': 'month'},
-          'seattle_temp']
-        }]
-    ds = DataPool(series_input) 
-    
+class BadChartOptionsTests(TestCase):
+
+    def setUp(self):
+        series_input = \
+          [{'options': {
+              'source': MonthlyWeatherByCity.objects.all()},
+            'terms': [
+              'month',
+              'boston_temp',
+              'houston_temp',
+              'new_york_temp']},
+           {'options': {
+              'source': MonthlyWeatherSeattle.objects.all()},
+            'terms': [
+              {'month_seattle': 'month'},
+              'seattle_temp']
+            }]
+        self.ds = DataPool(series_input)
+
     def test_options_missing(self):
         so_input = \
           [{'terms': {
@@ -1236,7 +1234,7 @@ class BadChartOptions(TestCase):
                  ['seattle_temp']}
             }]
         self.assertRaises(APIInputError, clean_cso, so_input, self.ds)
-    
+
     def test_options_wrong_type(self):
         so_input = \
           [{'options': 10,
@@ -1250,7 +1248,7 @@ class BadChartOptions(TestCase):
                  ['seattle_temp']}
             }]
         self.assertRaises(APIInputError, clean_cso, so_input, self.ds)
-        
+
     def test_terms_missing(self):
         so_input = \
           [{'options': {
@@ -1265,7 +1263,7 @@ class BadChartOptions(TestCase):
             'terms': 10
             }]
         self.assertRaises(APIInputError, clean_cso, so_input, self.ds)
-    
+
     def test_terms_a_list_not_a_dict(self):
         so_input = \
           [{'options': {
@@ -1274,7 +1272,7 @@ class BadChartOptions(TestCase):
               'month': ['new_york_temp']}]
             }]
         self.assertRaises(APIInputError, clean_cso, so_input, self.ds)
-        
+
     def test_terms_empty(self):
         so_input = \
           [{'options': {
@@ -1282,7 +1280,7 @@ class BadChartOptions(TestCase):
             'terms': {}
             }]
         self.assertRaises(APIInputError, clean_cso, so_input, self.ds)
-   
+
     def test_yterms_not_in_ds(self):
         so_input = \
           [{'options': {
@@ -1292,7 +1290,7 @@ class BadChartOptions(TestCase):
                  'foobar']}
             }]
         self.assertRaises(APIInputError, clean_cso, so_input, self.ds)
-    
+
     def test_xterms_not_in_ds(self):
         so_input = \
           [{'options': {
@@ -1302,7 +1300,7 @@ class BadChartOptions(TestCase):
                  'seattle_temp']}
             }]
         self.assertRaises(APIInputError, clean_cso, so_input, self.ds)
-    
+
     def test_x_and_y_not_in_same_table(self):
         so_input = \
           [{'options': {
@@ -1311,7 +1309,7 @@ class BadChartOptions(TestCase):
               'month_seattle':['new_york_temp']}
             }]
         self.assertRaises(APIInputError, clean_cso, so_input, self.ds)
-    
+
     def test_yterms_not_a_list(self):
         so_input = \
           [{'options': {
